# backend/openapi.yaml
openapi: 3.0.3
info:
  description: API para Sistema de Gestión de Condominios
  title: Casitas Teto API
  version: 1.0.0
servers:
- url: /
...
paths:
...
  /residentes:
    get:
      parameters:
      - description: ID del condominio
        in: query
        name: condominioId
        required: false
        schema:
          type: integer
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/_residentes_get_200_response"
          description: Lista de residentes
      summary: Listar residentes
      tags:
      - Residentes
    post:
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ResidenteInput"
        required: true
      responses:
        "201":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Residente"
          description: Residente registrado
      summary: Registrar residente
      tags:
      - Residentes
  /residentes/{residenteId}:
    get:
      parameters:
      - description: ID del residente
        in: path
        name: residenteId
        required: true
        schema:
          type: integer
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Residente"
          description: Detalles del residente
      summary: Obtener residente
      tags:
      - Residentes
    put:
      parameters:
      - description: ID del residente
        in: path
        name: residenteId
        required: true
        schema:
          type: integer
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ResidenteInput"
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Residente"
          description: Residente actualizado
      summary: Actualizar residente
      tags:
      - Residentes
...
  /multas:
    get:
      parameters:
      - description: ID del condominio
        in: query
        name: condominioId
        required: false
        schema:
          type: integer
      - in: query
        name: residenteId
        schema:
          type: integer
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/_multas_get_200_response"
          description: Lista de multas
      summary: Listar multas
      tags:
      - Multas
    post:
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MultaInput"
        required: true
      responses:
        "201":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Multa"
          description: Multa creada
      summary: Crear multa
      tags:
      - Multas
  /multas/{multaId}:
    get:
      parameters:
      - description: ID de la multa
        in: path
        name: multaId
        required: true
        schema:
          type: integer
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Multa"
          description: Detalles de la multa
      summary: Obtener multa
      tags:
      - Multas
    put:
      parameters:
      - description: ID de la multa
        in: path
        name: multaId
        required: true
        schema:
          type: integer
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MultaInput"
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Multa"
          description: Multa actualizada
      summary: Modificar multa
      tags:
      - Multas
  /multas/{multaId}/condonar:
    patch:
      parameters:
      - description: ID de la multa
        in: path
        name: multaId
        required: true
        schema:
          type: integer
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/_multas__multaId__condonar_patch_request"
      responses:
        "200":
          description: Multa condonada
      summary: Condonar multa
      tags:
      - Multas
...
  /pagos:
    get:
      parameters:
      - description: ID del condominio
        in: query
        name: condominioId
        required: false
        schema:
          type: integer
      - in: query
        name: residenteId
        schema:
          type: integer
      - in: query
        name: tipo
        schema:
          enum:
          - GASTO_COMUN
          - MULTA
          - RESERVA
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/_pagos_get_200_response"
          description: Lista de pagos
      summary: Listar pagos
      tags:
      - Pagos
    post:
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PagoInput"
        required: true
      responses:
        "201":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Pago"
          description: Pago registrado
      summary: Registrar pago
      tags:
      - Pagos
  /pagos/{pagoId}:
    get:
      parameters:
      - description: ID del pago
        in: path
        name: pagoId
        required: true
        schema:
          type: integer
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Pago"
          description: Detalles del pago
      summary: Obtener pago
      tags:
      - Pagos
  /pagos/{pagoId}/comprobante:
    get:
      parameters:
      - description: ID del pago
        in: path
        name: pagoId
        required: true
        schema:
          type: integer
      responses:
        "200":
          content:
            application/pdf:
              schema:
                format: binary
                type: string
          description: Comprobante de pago
      summary: Obtener comprobante de pago
      tags:
      - Pagos
    post:
      parameters:
      - description: ID del pago
        in: path
        name: pagoId
        required: true
        schema:
          type: integer
      requestBody:
        content:
          multipart/form-data:
            schema:
              $ref: "#/components/schemas/_pagos__pagoId__comprobante_post_request"
        required: true
      responses:
        "200":
          description: Comprobante adjuntado
      summary: Adjuntar comprobante de pago
      tags:
      - Pagos
...
components:
...
  schemas:
...
    Residente:
      example:
        rut: rut
        viviendaNumero: viviendaNumero
        fechaIngreso: 2000-01-23
        apellido: apellido
        condominioId: 1
        id: 0
        telefono: telefono
        usuarioId: 6
        nombre: nombre
        esPropietario: true
        email: email
        activo: true
      properties:
        id:
          title: id
          type: integer
        usuarioId:
          title: usuarioId
          type: integer
        condominioId:
          title: condominioId
          type: integer
        viviendaNumero:
          title: viviendaNumero
          type: string
        nombre:
          title: nombre
          type: string
        apellido:
          title: apellido
          type: string
        rut:
          title: rut
          type: string
        telefono:
          title: telefono
          type: string
        email:
          format: email
          title: email
          type: string
        esPropietario:
          title: esPropietario
          type: boolean
        fechaIngreso:
          format: date
          title: fechaIngreso
          type: string
        activo:
          title: activo
          type: boolean
      title: Residente
      type: object
    ResidenteInput:
      example:
        rut: rut
        viviendaNumero: viviendaNumero
        fechaIngreso: 2000-01-23
        apellido: apellido
        condominioId: 6
        telefono: telefono
        usuarioId: 0
        nombre: nombre
        esPropietario: true
        email: email
      properties:
        usuarioId:
          nullable: true
          title: usuarioId
          type: integer
        condominioId:
          title: condominioId
          type: integer
        viviendaNumero:
          title: viviendaNumero
          type: string
        nombre:
          title: nombre
          type: string
        apellido:
          title: apellido
          type: string
        rut:
          title: rut
          type: string
        telefono:
          title: telefono
          type: string
        email:
          format: email
          title: email
          type: string
        esPropietario:
          title: esPropietario
          type: boolean
        fechaIngreso:
          format: date
          title: fechaIngreso
          type: string
      required:
      - apellido
      - condominioId
      - email
      - esPropietario
      - nombre
      - rut
      - viviendaNumero
      title: ResidenteInput
      type: object
...
    Multa:
      example:
        descripcion: descripcion
        creadoPor: 5
        tipo: RETRASO_PAGO
        estado: PENDIENTE
        monto: 5.962133916683182
        motivoCondonacion: motivoCondonacion
        residenteId: 6
        condominioId: 1
        fechaEmision: 2000-01-23
        id: 0
        fechaPago: 2000-01-23T04:56:07.000+00:00
      properties:
        id:
          title: id
          type: integer
        residenteId:
          title: residenteId
          type: integer
        condominioId:
          title: condominioId
          type: integer
        tipo:
          enum:
          - RETRASO_PAGO
          - INFRAESTRUCTURA
          - RUIDO
          - MASCOTA
          - OTRO
          title: tipo
          type: string
        descripcion:
          title: descripcion
          type: string
        monto:
          format: decimal
          title: monto
          type: number
        estado:
          enum:
          - PENDIENTE
          - PAGADA
          - CONDONADA
          title: estado
          type: string
        fechaEmision:
          format: date
          title: fechaEmision
          type: string
        fechaPago:
          format: date-time
          nullable: true
          title: fechaPago
          type: string
        motivoCondonacion:
          nullable: true
          title: motivoCondonacion
          type: string
        creadoPor:
          title: creadoPor
          type: integer
      title: Multa
      type: object
    MultaInput:
      example:
        descripcion: descripcion
        tipo: RETRASO_PAGO
        monto: 0.14658129805029452
        residenteId: 0
        condominioId: 6
      properties:
        residenteId:
          title: residenteId
          type: integer
        condominioId:
          title: condominioId
          type: integer
        tipo:
          enum:
          - RETRASO_PAGO
          - INFRAESTRUCTURA
          - RUIDO
          - MASCOTA
          - OTRO
          title: tipo
          type: string
        descripcion:
          title: descripcion
          type: string
        monto:
          format: decimal
          minimum: 0
          title: monto
          type: number
      required:
      - condominioId
      - descripcion
      - monto
      - residenteId
      - tipo
      title: MultaInput
      type: object
...
    Pago:
      example:
        referenciaId: 5
        metodoPago: TRANSFERENCIA
        tipo: GASTO_COMUN
        monto: 5.637376656633329
        numeroTransaccion: numeroTransaccion
        comprobanteUrl: https://openapi-generator.tech
        residenteId: 1
        condominioId: 6
        registradoPor: 2
        id: 0
        estadoPago: PENDIENTE
        fechaPago: 2000-01-23T04:56:07.000+00:00
      properties:
        id:
          title: id
          type: integer
        condominioId:
          title: condominioId
          type: integer
        residenteId:
          title: residenteId
          type: integer
        tipo:
          enum:
          - GASTO_COMUN
          - MULTA
          - RESERVA
          title: tipo
          type: string
        referenciaId:
          title: referenciaId
          type: integer
        monto:
          format: decimal
          title: monto
          type: number
        metodoPago:
          enum:
          - TRANSFERENCIA
          - TARJETA
          - EFECTIVO
          - WEBPAY
          - KHIPU
          title: metodoPago
          type: string
        estadoPago:
          enum:
          - PENDIENTE
          - APROBADO
          - RECHAZADO
          - REVERSADO
          title: estadoPago
          type: string
        numeroTransaccion:
          nullable: true
          title: numeroTransaccion
          type: string
        fechaPago:
          format: date-time
          title: fechaPago
          type: string
        comprobanteUrl:
          format: uri
          nullable: true
          title: comprobanteUrl
          type: string
        registradoPor:
          title: registradoPor
          type: integer
      title: Pago
      type: object
    PagoInput:
      example:
        referenciaId: 1
        metodoPago: TRANSFERENCIA
        tipo: GASTO_COMUN
        monto: 0.5962133916683182
        numeroTransaccion: numeroTransaccion
        residenteId: 6
        condominioId: 0
      properties:
        condominioId:
          title: condominioId
          type: integer
        residenteId:
          title: residenteId
          type: integer
        tipo:
          enum:
          - GASTO_COMUN
          - MULTA
          - RESERVA
          title: tipo
          type: string
        referenciaId:
          title: referenciaId
          type: integer
        monto:
          format: decimal
          minimum: 0
          title: monto
          type: number
        metodoPago:
          enum:
          - TRANSFERENCIA
          - TARJETA
          - EFECTIVO
          - WEBPAY
          - KHIPU
          title: metodoPago
          type: string
        numeroTransaccion:
          title: numeroTransaccion
          type: string
      required:
      - condominioId
      - metodoPago
      - monto
      - referenciaId
      - residenteId
      - tipo
      title: PagoInput
      type: object
...
    _residentes_get_200_response:
      example:
        data:
        - rut: rut
          viviendaNumero: viviendaNumero
          fechaIngreso: 2000-01-23
          apellido: apellido
          condominioId: 1
          id: 0
          telefono: telefono
          usuarioId: 6
          nombre: nombre
          esPropietario: true
          email: email
          activo: true
        - rut: rut
          viviendaNumero: viviendaNumero
          fechaIngreso: 2000-01-23
          apellido: apellido
          condominioId: 1
          id: 0
          telefono: telefono
          usuarioId: 6
          nombre: nombre
          esPropietario: true
          email: email
          activo: true
      properties:
        data:
          items:
            $ref: "#/components/schemas/Residente"
          title: data
          type: array
      title: _residentes_get_200_response
      type: object
...
    _multas_get_200_response:
      example:
        data:
        - descripcion: descripcion
          creadoPor: 5
          tipo: RETRASO_PAGO
          estado: PENDIENTE
          monto: 5.962133916683182
          motivoCondonacion: motivoCondonacion
          residenteId: 6
          condominioId: 1
          fechaEmision: 2000-01-23
          id: 0
          fechaPago: 2000-01-23T04:56:07.000+00:00
        - descripcion: descripcion
          creadoPor: 5
          tipo: RETRASO_PAGO
          estado: PENDIENTE
          monto: 5.962133916683182
          motivoCondonacion: motivoCondonacion
          residenteId: 6
          condominioId: 1
          fechaEmision: 2000-01-23
          id: 0
          fechaPago: 2000-01-23T04:56:07.000+00:00
      properties:
        data:
          items:
            $ref: "#/components/schemas/Multa"
          title: data
          type: array
      title: _multas_get_200_response
      type: object
    _multas__multaId__condonar_patch_request:
      properties:
        motivo:
          title: motivo
          type: string
      title: _multas__multaId__condonar_patch_request
      type: object
...
    _pagos_get_200_response:
      example:
        data:
        - referenciaId: 5
          metodoPago: TRANSFERENCIA
          tipo: GASTO_COMUN
          monto: 5.637376656633329
          numeroTransaccion: numeroTransaccion
          comprobanteUrl: https://openapi-generator.tech
          residenteId: 1
          condominioId: 6
          registradoPor: 2
          id: 0
          estadoPago: PENDIENTE
          fechaPago: 2000-01-23T04:56:07.000+00:00
        - referenciaId: 5
          metodoPago: TRANSFERENCIA
          tipo: GASTO_COMUN
          monto: 5.637376656633329
          numeroTransaccion: numeroTransaccion
          comprobanteUrl: https://openapi-generator.tech
          residenteId: 1
          condominioId: 6
          registradoPor: 2
          id: 0
          estadoPago: PENDIENTE
          fechaPago: 2000-01-23T04:56:07.000+00:00
      properties:
        data:
          items:
            $ref: "#/components/schemas/Pago"
          title: data
          type: array
      title: _pagos_get_200_response
      type: object
    _pagos__pagoId__comprobante_post_request:
      properties:
        archivo:
          format: binary
          type: string
      type: object
...